refactor(workshop): get-stat moved into workshop
