--- libs/deno-utils/mod.ts
@@ -1,1 +1,0 @@
-export * from './get-stat.ts'
