+++ libs/workshop/setup.ts
@@ -7,7 +7,7 @@ import {
   writePlaywrightConfig,
   writeYarnrc,
 } from './write/mod.ts'
-import { getStat } from '../deno-utils/mod.ts'
+import { getStat } from './get-stat.ts'
 
 export const setup = async ({
   credentials,
