+++ libs/workshop/write/write-package.ts
@@ -1,5 +1,5 @@
 import { packageJson } from '../templates/mod.ts'
-import { getStat } from '../../deno-utils/mod.ts'
+import { getStat } from '../get-stat.ts'
 
 export const writePackage = async (path: string) => {
   const exist = await getStat(path)
