+++ libs/workshop/write/write-playwright-config.ts
@@ -1,5 +1,5 @@
 import { playwrightConfig } from '../templates/mod.ts'
-import { getStat } from '../../deno-utils/mod.ts'
+import { getStat } from '../get-stat.ts'
 
 type WritePlaywrightConfig = (args: {
   path: string
