+++ libs/workshop/write/write-yarnrc.ts
@@ -1,5 +1,5 @@
 import { yarnrc } from '../templates/mod.ts'
-import { getStat } from '../../deno-utils/mod.ts'
+import { getStat } from '../get-stat.ts'
 
 export const writeYarnrc = async (path: string) => {
   const exist = await getStat(path)
